% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Scrape Sagarin ratings}
\usage{
scrape(
  url = "https://sagarin.usatoday.com/2023-2/college-basketball-team-ratings-2022-23/"
)
}
\arguments{
\item{url}{a character indicating the url to scrape}
}
\value{
a data.frame containing all the scraped data
}
\description{
Scrape Sagarin ratings from the USA Today website.
}
